# server/openapi.yaml
openapi: 3.0.3
info:
  title: Huuk API (by teenueng)
  version: 0.1.0
  description: >-
    REST API for Huuk by teenueng (Vue 3 frontend in `app/`,
    Rust/Axum backend in `server/`). All request and response bodies are JSON
    with camelCase field names; errors use the envelope
    `{ "error": "message" }`.

    Authentication: send `Authorization: Bearer <token>` from
    `POST /api/auth/login`. Auth is only enforced when `authRequired` is true
    (production default); in demo mode (`DEMO_MODE=1`) reads and writes are
    open and the Bearer token is optional. In production every `GET` also
    requires a live token, except liveness (`/api/health`, `/api/ready`), the
    redacted `GET /api/setup`, the public delivery API (`/api/public/*`), the
    brand CI (`/api/brand`, `/api/brand/fonts`, font files) and the OAuth/docs
    routes; every mutating operation is gated by the permission named in its
    `x-permission` field.

    Content items (`/api/content`) implement an editorial workflow with
    optimistic concurrency: send the `version` you last read on every
    update/publish/schedule/restore call and expect `409` when someone else
    saved in between.

    Interactive docs are served by the app itself: `GET /api/docs` (Swagger UI)
    and `GET /api/openapi.yaml` (this document).
servers:
  - url: /
    description: Same origin (the server also serves the built SPA when STATIC_DIR is set)
  - url: http://localhost:8787
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/workspaces:
    get:
      operationId: listWorkspaces
      tags: [Workspaces]
      summary: List workspaces
      description: >-
        Workspaces separate brands: each owns its own brand identity and social
        connections. Every request targets one via the `X-Workspace-Id` header
        (falling back to the first workspace).
      responses:
        '200':
          description: All workspaces with connection counts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WorkspaceSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createWorkspace
      tags: [Workspaces]
      summary: Create a workspace
      x-permission: platforms.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
      responses:
        '200':
          description: The created workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/workspaces/{id}:
    patch:
      operationId: renameWorkspace
      tags: [Workspaces]
      summary: Rename a workspace
      x-permission: platforms.manage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
      responses:
        '200':
          description: The renamed workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteWorkspace
      tags: [Workspaces]
      summary: Delete a workspace
      description: >-
        Drops the workspace with everything scoped to it (connections, provider
        tokens, pending OAuth state). The last workspace cannot be deleted.
      x-permission: platforms.manage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deletion result with the fallback workspace id.
          content:
            application/json:
              schema:
                type: object
                required: [deleted, fallback]
                properties:
                  deleted: { type: string }
                  fallback: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/workspaces/{id}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [Workspaces]
      summary: List workspace members
      description: >-
        The owner plus the registered accounts the owner added. Only the owner
        can read or change the list.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The owner and the member list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceMembers' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: addWorkspaceMember
      tags: [Workspaces]
      summary: Add a workspace member
      description: >-
        Adds a registered account by email. The account must already exist; the
        owner and existing members are rejected.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: The updated owner and member list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceMembers' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/workspaces/{id}/members/{email}:
    delete:
      operationId: removeWorkspaceMember
      tags: [Workspaces]
      summary: Remove a workspace member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: email
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The updated owner and member list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceMembers' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token from `POST /api/auth/login`. Optional while
        `authRequired` is false (demo mode); required for mutations in
        production.
...
  responses:
    BadRequest:
      description: Invalid payload or failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, unknown or expired bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The signed-in role lacks the required permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Unknown id or route.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Locked by another user, duplicate email/name, or version conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: post is locked by Studio Owner }
...
    WorkspaceSummary:
      type: object
      required: [id, name, created, connected, total, isOwner]
      properties:
        id: { type: string }
        name: { type: string }
        created: { type: string }
        connected: { type: integer, description: Platforms currently connected. }
        total: { type: integer, description: Platform slots in the workspace. }
        isOwner: { type: boolean, description: Whether the requesting account owns the workspace. }
    WorkspaceMembers:
      type: object
      required: [owner, members]
      properties:
        owner:
          type: object
          required: [email, name]
          properties:
            email: { type: string }
            name: { type: string }
        members:
          type: array
          items:
            type: object
            required: [email, name, role]
            properties:
              email: { type: string }
              name: { type: string }
              role: { type: string }
...